Extend list of supported estimators for shadow models
